chore: split transaction parser - update project structure
