+++ apps/transaction-parser/src/payobills/mod.rs
@@ -1,0 +1,1 @@
+pub mod transaction_parser;
